move precompiles to sub package
