+++ precompile/contracts/feemanager/module.go
@@ -1,4 +1,4 @@
-// (c) 2023, Ava Labs, Inc. All rights reserved.
+// (c) 2019-2020, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
 package feemanager
@@ -8,12 +8,34 @@ import (
 
 	"github.com/ava-labs/subnet-evm/precompile/config"
 	"github.com/ava-labs/subnet-evm/precompile/contract"
+	"github.com/ethereum/go-ethereum/common"
 )
 
-type Executor struct{}
+var _ contract.Module = Module{}
+
+// ConfigKey is the key used in json config files to specify this precompile config.
+// must be unique across all precompiles.
+const ConfigKey = "feeManagerConfig"
+
+var ContractAddress = common.HexToAddress("0x0200000000000000000000000000000000000003")
+
+type Module struct{}
+
+func (Module) Key() string {
+	return ConfigKey
+}
+
+// Address returns the address of the fee manager.
+func (Module) Address() common.Address {
+	return ContractAddress
+}
+
+func (Module) NewConfig() config.Config {
+	return &FeeManagerConfig{}
+}
 
 // Configure configures [state] with the desired admins based on [configIface].
-func (Executor) Configure(chainConfig contract.ChainConfig, cfg config.Config, state contract.StateDB, blockContext contract.BlockContext) error {
+func (Module) Configure(chainConfig contract.ChainConfig, cfg config.Config, state contract.StateDB, blockContext contract.BlockContext) error {
 	config, ok := cfg.(*FeeManagerConfig)
 	if !ok {
 		return fmt.Errorf("incorrect config %T: %v", config, config)
@@ -33,6 +55,6 @@ func (Executor) Configure(chainConfig contract.ChainConfig, cfg config.Config, s
 	return config.AllowListConfig.Configure(state, ContractAddress)
 }
 
-func (Executor) Contract() contract.StatefulPrecompiledContract {
+func (Module) Contract() contract.StatefulPrecompiledContract {
 	return FeeManagerPrecompile
 }
