+++ precompile/contracts/nativeminter/contract.go
