+++ precompile/contracts/rewardmanager/config.go
