+++ precompile/contracts/rewardmanager/config_test.go
