+++ precompile/contracts/rewardmanager/contract.go
