+++ precompile/contracts/rewardmanager/reward_manager_test.go
@@ -298,7 +298,7 @@ func TestRewardManagerRun(t *testing.T) {
 			accesibleState := contract.NewMockAccessibleState(state, blockContext, snow.DefaultContextTest())
 
 			if test.config != nil {
-				Executor{}.Configure(nil, test.config, state, blockContext)
+				Module{}.Configure(nil, test.config, state, blockContext)
 			}
 			ret, remainingGas, err := RewardManagerPrecompile.Run(accesibleState, test.caller, ContractAddress, test.input(), test.suppliedGas, test.readOnly)
 			if len(test.expectedErr) != 0 {
