+++ precompile/contracts/txallowlist/module.go
@@ -4,10 +4,14 @@
 package txallowlist
 
 import (
+	"fmt"
+
 	"github.com/ava-labs/subnet-evm/precompile/config"
 	"github.com/ava-labs/subnet-evm/precompile/contract"
 	"github.com/ethereum/go-ethereum/common"
 )
+
+var _ contract.Module = Module{}
 
 // ConfigKey is the key used in json config files to specify this precompile config.
 // must be unique across all precompiles.
@@ -30,6 +34,15 @@ func (Module) NewConfig() config.Config {
 	return &TxAllowListConfig{}
 }
 
-func (Module) Executor() contract.Execution {
-	return &Executor{}
+// Configure configures [state] with the desired admins based on [cfg].
+func (Module) Configure(chainConfig contract.ChainConfig, cfg config.Config, state contract.StateDB, _ contract.BlockContext) error {
+	config, ok := cfg.(*TxAllowListConfig)
+	if !ok {
+		return fmt.Errorf("incorrect config %T: %v", config, config)
+	}
+	return config.AllowListConfig.Configure(state, ContractAddress)
+}
+
+func (Module) Contract() contract.StatefulPrecompiledContract {
+	return TxAllowListPrecompile
 }
