+++ precompile/registry/registerer.go
@@ -1,7 +1,7 @@
 // (c) 2023, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package modules
+package registry
 
 import (
 	"fmt"
