--- precompile/txallowlist/execution.go
@@ -1,26 +1,0 @@
-// (c) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package txallowlist
-
-import (
-	"fmt"
-
-	"github.com/ava-labs/subnet-evm/precompile/config"
-	"github.com/ava-labs/subnet-evm/precompile/contract"
-)
-
-type Executor struct{}
-
-// Configure configures [state] with the desired admins based on [cfg].
-func (Executor) Configure(chainConfig contract.ChainConfig, cfg config.Config, state contract.StateDB, _ contract.BlockContext) error {
-	config, ok := cfg.(*TxAllowListConfig)
-	if !ok {
-		return fmt.Errorf("incorrect config %T: %v", config, config)
-	}
-	return config.AllowListConfig.Configure(state, ContractAddress)
-}
-
-func (Executor) Contract() contract.StatefulPrecompiledContract {
-	return TxAllowListPrecompile
-}
